+++ src/reports.rs
@@ -1,6 +1,9 @@
 use crate::workspace::Workspace;
-use crate::{print_lines, OutputFormat};
+use crate::OutputFormat;
 use anyhow::Result;
+use std::collections::HashSet;
+use tabled::builder::Builder;
+use tabled::settings::Style;
 
 pub fn display_affected_files(workspace: &Workspace, format: &OutputFormat) -> Result<()> {
     let file_paths = workspace.affected_files()?;
@@ -25,5 +28,48 @@ pub fn display_affected_projects(workspace: &Workspace, format: &OutputFormat) -
 
     print_lines(&projects, format, "Project")?;
 
+    Ok(())
+}
+
+pub fn display_tasks(workspace: &Workspace, format: &OutputFormat) -> Result<()> {
+    let tasks = workspace.tasks();
+
+    if tasks.is_empty() {
+        println!("No tasks defined");
+        return Ok(());
+    }
+
+    for task in tasks {
+        println!("{}", task);
+    }
+
+    Ok(())
+}
+
+pub fn print_lines(lines: &HashSet<String>, format: &OutputFormat, header: &str) -> Result<()> {
+    match format {
+        OutputFormat::Json => {
+            let json_output = serde_json::to_string_pretty(&lines)?;
+            println!("{}", json_output);
+        }
+        OutputFormat::Table => {
+            let mut builder = Builder::default();
+            builder.push_record(["#", header]);
+
+            for (index, line) in lines.iter().enumerate().map(|(i, line)| (i + 1, line)) {
+                builder.push_record([&index.to_string(), line]);
+            }
+
+            let mut table = builder.build();
+            table.with(Style::modern());
+
+            println!("{}", table);
+        }
+        _ => {
+            for line in lines {
+                println!("{}", line);
+            }
+        }
+    }
     Ok(())
 }
